# event_go/internal/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soulmarker Event API",
    "version": "1.0.0",
    "description": "Versioned API contract for authentication, organization authorization, organizers, events, registrations, admissions, checkins, discussions and tickets. Legacy /api routes remain available for one compatibility cycle."
  },
  "servers": [{ "url": "/api/v1" }],
...
  "paths": {
...
    "/me/admissions": {
      "get": {
        "operationId": "listMyAdmissions",
        "tags": ["Me", "Admissions"],
        "summary": "List current and revoked free-event admission credentials",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/me/activities": {
      "get": {
        "operationId": "listMyActivities",
        "tags": ["Me", "Admissions"],
        "summary": "List a stable unified timeline of registrations and admissions",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/organizations/{organizationId}/members": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "listOrganizationMembers",
        "tags": ["Organizations"],
        "summary": "List members visible to the current tenant principal",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/organizations/{organizationId}/owner-transfer": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "post": {
        "operationId": "transferOrganizationOwnership",
        "tags": ["Organizations"],
        "summary": "Transfer organization ownership to another active member",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/TransferOrganizationOwnerRequest" },
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/organizations/{organizationId}/invitations": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "listOrganizationInvitations",
        "tags": ["Organizations"],
        "summary": "List organization invitations without secret token material",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "operationId": "createOrganizationInvitation",
        "tags": ["Organizations"],
        "summary": "Invite a user to the organization and deliver a one-time link",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateOrganizationInvitationRequest" },
        "responses": {
          "201": { "$ref": "#/components/responses/Created" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/organizations/{organizationId}/audits": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "get": {
        "operationId": "listOrganizationAudits",
        "tags": ["Organizations", "Audit"],
        "summary": "List tenant-scoped organization audit logs",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/PageSize" }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/organizations/{organizationId}/events": {
      "parameters": [{ "$ref": "#/components/parameters/OrganizationID" }],
      "post": {
        "operationId": "createOrganizationEvent", "tags": ["Organizations", "Events"],
        "summary": "Create an event inside the authorized tenant scope", "security": [{ "BearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateEventRequest" },
        "responses": { "201": { "$ref": "#/components/responses/Created" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "get": {
        "operationId": "listOrganizationEvents", "tags": ["Organizations", "Events"],
        "summary": "List only events owned by the authorized tenant", "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/PageSize" }],
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
    "/events": {
      "post": {
        "operationId": "createEvent",
        "tags": ["Events"],
        "summary": "Create an event",
        "security": [{ "AdminToken": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CreateEventRequest" },
        "responses": { "201": { "$ref": "#/components/responses/Created" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      },
      "get": {
        "operationId": "listEvents",
        "tags": ["Events"],
        "summary": "List and filter events",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "published", "cancelled", "ended"] } },
          { "name": "price_type", "in": "query", "schema": { "type": "string", "enum": ["free", "paid"] } },
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "organizer_id", "in": "query", "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": { "200": { "$ref": "#/components/responses/OK" } }
      }
    },
    "/events/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/EventID" }],
      "get": {
        "operationId": "getEvent",
        "tags": ["Events"],
        "summary": "Get an event",
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "put": {
        "operationId": "updateEvent",
        "tags": ["Events"],
        "summary": "Update an event",
        "security": [{ "AdminToken": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/UpdateEventRequest" },
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "delete": {
        "operationId": "deleteEvent",
        "tags": ["Events"],
        "summary": "Delete an event and dependent records",
        "security": [{ "AdminToken": [] }],
        "responses": { "200": { "$ref": "#/components/responses/OK" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
    "/events/{id}/admission": {
      "parameters": [{ "$ref": "#/components/parameters/EventID" }],
      "get": {
        "operationId": "getMyAdmission",
        "tags": ["Admissions"],
        "summary": "Get the current user's admission credential for an event",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "AdminToken": { "type": "apiKey", "in": "header", "name": "X-Admin-Token", "description": "Platform administrator credential; it does not grant an organization membership or tenant capability." }
    },
    "parameters": {
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "PageSize": { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
      "OrganizationID": { "name": "organizationId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
      "EventID": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
...
    },
    "requestBodies": {
...
      "CreateOrganizationInvitationRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrganizationInvitationRequest" } } } },
...
      "TransferOrganizationOwnerRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransferOrganizationOwnerRequest" } } } },
...
      "CreateEventRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateEventRequest" } } } },
      "UpdateEventRequest": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateEventRequest" } } } },
...
    },
    "responses": {
      "OK": { "description": "Successful response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/APIResponse" } } } },
      "Created": { "description": "Resource created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/APIResponse" } } } },
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Unauthorized": { "description": "Authentication failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Forbidden": { "description": "Authenticated but not allowed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "NotFound": { "description": "Resource not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Conflict": { "description": "Business state conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
    },
    "schemas": {
      "APIResponse": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "integer" },
          "error_code": { "type": "string", "enum": ["ADMIN_AUTH_INVALID", "ADMISSION_ALREADY_CHECKED_IN", "ADMISSION_NOT_FOUND", "ADMISSION_REVOKED", "API_ROUTE_NOT_FOUND", "CANCELLATION_DEADLINE_EXCEEDED", "CONTENT_ALREADY_REMOVED", "CONTENT_ALREADY_VISIBLE", "CONTENT_REPORT_ALREADY_RESOLVED", "CONTENT_REPORT_NOT_ALLOWED", "CONTENT_REPORT_NOT_FOUND", "EVENT_CAPACITY_FULL", "EVENT_HAS_ADMISSIONS", "EVENT_NOT_FOUND", "EVENT_NOT_PUBLISHED", "INTERNAL_ERROR", "INVALID_CREDENTIALS", "INVALID_JSON", "METHOD_NOT_ALLOWED", "NOTIFICATION_NOT_FOUND", "ORGANIZATION_ACCESS_DENIED", "ORGANIZATION_INVITATION_EXISTS", "ORGANIZATION_INVITATION_INVALID", "ORGANIZATION_MEMBER_CHANGE_DENIED", "ORGANIZATION_MEMBER_EXISTS", "ORGANIZATION_MEMBER_NOT_FOUND", "ORGANIZATION_OWNER_TRANSFER_DENIED", "ORGANIZATION_SLUG_IN_USE", "ORGANIZER_NOT_FOUND", "PARTICIPATION_REQUIRED", "PASSWORD_RESET_TOKEN_INVALID", "POST_NOT_FOUND", "RECOVERY_EMAIL_ALREADY_BOUND", "RECOVERY_EMAIL_IN_USE", "RECOVERY_EMAIL_RATE_LIMITED", "RECOVERY_EMAIL_TOKEN_INVALID", "REGISTRATION_DUPLICATE", "REGISTRATION_NOT_FOUND", "REPLY_NOT_FOUND", "REQUEST_TOO_LARGE", "TICKET_NOT_FOUND", "TICKET_SOLD_OUT", "USER_ALREADY_EXISTS", "USER_AUTH_REQUIRED", "USER_TOKEN_INVALID", "VALIDATION_ERROR"] },
          "message": { "type": "string" },
          "data": {},
          "total": { "type": "integer", "minimum": 0 },
          "page": { "type": "integer", "minimum": 1 },
          "page_size": { "type": "integer", "minimum": 1, "maximum": 100 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "error_code", "message"],
        "properties": {
          "code": { "type": "integer", "minimum": 400, "maximum": 599 },
          "error_code": { "type": "string", "enum": ["ADMIN_AUTH_INVALID", "ADMISSION_ALREADY_CHECKED_IN", "ADMISSION_NOT_FOUND", "ADMISSION_REVOKED", "API_ROUTE_NOT_FOUND", "CANCELLATION_DEADLINE_EXCEEDED", "CONTENT_ALREADY_REMOVED", "CONTENT_ALREADY_VISIBLE", "CONTENT_REPORT_ALREADY_RESOLVED", "CONTENT_REPORT_NOT_ALLOWED", "CONTENT_REPORT_NOT_FOUND", "EVENT_CAPACITY_FULL", "EVENT_HAS_ADMISSIONS", "EVENT_NOT_FOUND", "EVENT_NOT_PUBLISHED", "INTERNAL_ERROR", "INVALID_CREDENTIALS", "INVALID_JSON", "METHOD_NOT_ALLOWED", "NOTIFICATION_NOT_FOUND", "ORGANIZATION_ACCESS_DENIED", "ORGANIZATION_INVITATION_EXISTS", "ORGANIZATION_INVITATION_INVALID", "ORGANIZATION_MEMBER_CHANGE_DENIED", "ORGANIZATION_MEMBER_EXISTS", "ORGANIZATION_MEMBER_NOT_FOUND", "ORGANIZATION_OWNER_TRANSFER_DENIED", "ORGANIZATION_SLUG_IN_USE", "ORGANIZER_NOT_FOUND", "PARTICIPATION_REQUIRED", "PASSWORD_RESET_TOKEN_INVALID", "POST_NOT_FOUND", "RECOVERY_EMAIL_ALREADY_BOUND", "RECOVERY_EMAIL_IN_USE", "RECOVERY_EMAIL_RATE_LIMITED", "RECOVERY_EMAIL_TOKEN_INVALID", "REGISTRATION_DUPLICATE", "REGISTRATION_NOT_FOUND", "REPLY_NOT_FOUND", "REQUEST_TOO_LARGE", "TICKET_NOT_FOUND", "TICKET_SOLD_OUT", "USER_ALREADY_EXISTS", "USER_AUTH_REQUIRED", "USER_TOKEN_INVALID", "VALIDATION_ERROR"] },
          "message": { "type": "string" }
        }
      },
...
      "CreateOrganizationInvitationRequest": { "type": "object", "additionalProperties": false, "required": ["email", "role"], "properties": { "email": { "type": "string", "format": "email" }, "role": { "type": "string", "enum": ["admin", "editor", "checker", "finance"] } } },
...
      "TransferOrganizationOwnerRequest": { "type": "object", "additionalProperties": false, "required": ["member_id"], "properties": { "member_id": { "type": "integer", "format": "int64", "minimum": 1 } } },
...
      "CreateEventRequest": { "type": "object", "additionalProperties": false, "required": ["organizer_id", "title", "event_time", "location", "capacity", "price"], "properties": { "organizer_id": { "type": "integer", "format": "int64", "minimum": 1 }, "title": { "type": "string" }, "description": { "type": "string" }, "cover_url": { "type": "string", "format": "uri-reference" }, "event_time": { "type": "string", "format": "date-time" }, "location": { "type": "string" }, "capacity": { "type": "integer", "minimum": 1 }, "price": { "type": "number", "minimum": 0 } } },
      "UpdateEventRequest": { "type": "object", "additionalProperties": false, "properties": { "organizer_id": { "type": "integer", "format": "int64", "minimum": 1 }, "title": { "type": "string" }, "description": { "type": "string" }, "cover_url": { "type": "string", "format": "uri-reference" }, "event_time": { "type": "string", "format": "date-time" }, "location": { "type": "string" }, "capacity": { "type": "integer", "minimum": 1 }, "price": { "type": "number", "minimum": 0 }, "status": { "type": "string", "enum": ["draft", "published", "cancelled", "ended"] } } },
...
    }
  }
}